% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/services_lookup.R
\name{services_lookup}
\alias{services_lookup}
\title{Services}
\usage{
services_lookup(Services)
}
\description{
Creates a dataframe to match JourneyPatternRef with JourneyPatternid
For some files there are multiple JourneyPatternRefs for each JourneyPatternid so they are iterated over
}
