% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stop_points.R
\name{stop_point_load}
\alias{stop_point_load}
\title{A basic function to convert the stop point list to a data frame}
\usage{
stop_point_load(StopPoints)
}
\description{
A basic function to convert the stop point list to a data frame
}
