% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timetables_from_vehicles_journeys.R
\name{timetable}
\alias{timetable}
\title{timetable}
\usage{
timetable(vehicles, journeys, services)
}
\description{
Construct full timings
}
\details{
The timetable constructed is a list of lists with the first three elements containing information
and the fourth element containing the actual timetable
}
