% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transxchange_xml_load.R
\name{transxchange_xml_load}
\alias{transxchange_xml_load}
\title{A basic function to load an xml file and split it into
the component lists
Requires XML}
\usage{
transxchange_xml_load(file_name)
}
\description{
A basic function to load an xml file and split it into
the component lists
Requires XML
}
