% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VehicleJourneys_to_dataframe.R
\name{VehicleJourneys_to_dataframe}
\alias{VehicleJourneys_to_dataframe}
\title{VehicleJourneys_to_dataframe}
\usage{
VehicleJourneys_to_dataframe(VehicleJourneys)
}
\description{
A basic function to get departure time and operating times for each vehicle journey
}
\details{
Requires XML
}
