% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stop_points.R
\name{stop_point_load}
\alias{stop_point_load}
\title{stop_point_load}
\usage{
stop_point_load(StopPoints)
}
\description{
A basic function to convert the stop point list to a data frame.
GPS data can be derived from the master Naptan list
}
